Refactor feature code into separate `cesium.features` module
