+++ cesium/__init__.py
@@ -1,6 +1,6 @@
 """Machine Learning Time-Series Platform (cesium)
 
-See http://cesium.mlfor more information.
+See http://cesium.ml for more information.
 """
 
 from .version import version as __version__
