+++ cesium/features/_lomb_scargle.h
