+++ cesium/features/_lomb_scargle.pxd
