+++ cesium/setup.py
@@ -2,7 +2,7 @@ def configuration(parent_package='', top_path=None):
     from numpy.distutils.misc_util import Configuration
 
     config = Configuration('cesium', parent_package, top_path)
-    config.add_subpackage('science_features')
+    config.add_subpackage('features')
     config.add_data_dir('tests')
     config.add_data_dir('tests/data')
     return config
